controllers/repositories/repositories_controller_test.go
 package repositories

//import (
//	"encoding/json"
//	"github.com/dembygenesis/go-rest-industry-standard/src/api/clients/restclient"
//	"github.com/dembygenesis/go-rest-industry-standard/src/api/domain/repositories"
//	"github.com/dembygenesis/go-rest-industry-standard/src/api/utils/errors"
//	"github.com/dembygenesis/go-rest-industry-standard/src/api/utils/test_utils"
//	"github.com/stretchr/testify/assert"
//	"io/ioutil"
//	"net/http"
//	"net/http/httptest"
//	"os"
//	"strings"
//	"testing"
//)
//
//func TestMain(m *testing.M) {
//	restclient.StartMockups()
//	os.Exit(m.Run())
//}
//
//func TestCreateRepoInvalidJsonRequest(t *testing.T) {
//	response := httptest.NewRecorder()
//	request, _ := http.NewRequest(http.MethodPost, "/repositories", strings.NewReader(``))
//	c := test_utils.GetMockedContext(request, response)
//
//	CreateRepo(c)
//
//	assert.EqualValues(t, http.StatusBadRequest, response.Code)
//
//	apiErr, err := errors.NewApiErrorFromBytes(response.Body.Bytes())
//	assert.Nil(t, err)
//	assert.NotNil(t, apiErr)
//	assert.EqualValues(t, http.StatusBadRequest, apiErr.Status())
//	assert.EqualValues(t, "invalid json body", apiErr.Message())
//}
//
//func TestCreateRepoErrorFromGithub(t *testing.T) {
//	restclient.FlushMockups()
//	restclient.AddMockup(restclient.Mock{
//		Url:        "https://api.github.com/user/repos",
//		HttpMethod: http.MethodPost,
//		Response: &http.Response{
//			StatusCode: http.StatusUnauthorized,
//			Body:       ioutil.NopCloser(strings.NewReader(`{"status": 401,"message": "Requires authentication"}`)),
//		},
//	})
//
//	response := httptest.NewRecorder()
//	request, _ := http.NewRequest(http.MethodPost, "/repositories", strings.NewReader(`{"name": "testing"}`))
//	c := test_utils.GetMockedContext(request, response)
//
//	// Execute gin controller function
//	CreateRepo(c)
//
//	assert.EqualValues(t, http.StatusUnauthorized, response.Code)
//
//	apiErr, err := errors.NewApiErrorFromBytes(response.Body.Bytes())
//	assert.Nil(t, err)
//	assert.NotNil(t, apiErr)
//	assert.EqualValues(t, http.StatusUnauthorized, apiErr.Status())
//	assert.EqualValues(t, "Requires authentication", apiErr.Message())
//}
//
//func TestCreateRepoNoError(t *testing.T) {
//	restclient.FlushMockups()
//	restclient.AddMockup(restclient.Mock{
//		Url:        "https://api.github.com/user/repos",
//		HttpMethod: http.MethodPost,
//		Response: &http.Response{
//			StatusCode: http.StatusCreated,
//			Body:       ioutil.NopCloser(strings.NewReader(`{"id": 123}`)),
//		},
//	})
//
//	response := httptest.NewRecorder()
//	request, _ := http.NewRequest(http.MethodPost, "/repositories", strings.NewReader(`{"name": "testing"}`))
//	c := test_utils.GetMockedContext(request, response)
//
//	// Execute gin controller function
//	CreateRepo(c)
//
//	assert.EqualValues(t, http.StatusCreated, response.Code)
//
//	var result repositories.CreateRepoResponse
//	err := json.Unmarshal(response.Body.Bytes(), &result)
//
//	assert.Nil(t, err)
//	assert.EqualValues(t, 123, result.Id)
//	assert.EqualValues(t, "", result.Name)
//	assert.EqualValues(t, "", result.Owner)
//}

